New SQL Scripts
Our first SQL scripts were creating and populating a database created via. "Forward Engineer" from the EER Diagram, but we've discovered that this method applies some bad constrains to our table and key relationships which was preventing us from manipulating the parent tables. The modifications were made mainly to the "SQL.test.script" which is responsible for creating the database with the correct constrains and variable properties that we need in order for Java Spring to comprehend the functionality. The database created from this script is still following the structure of the EER Diagram initially uploaded, therefore the Structure of the database remained the same.